UPDATE THRESHOLD IN LAMBDA FN
Realized distance had to be allowed to get to 1e-15 if the deltaS is often getting to 1e-6 in the iterative final. Then lambda threshold in main loop is actually used.

Also allowed for flexability on highly coplanar intersections - sometimes this is the case on TIR and we still need lambda value.